pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "figmaclaw"
version = "0.1.86"
description = "Figma → git semantic design memory for AI agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "pydantic>=2.7",
    "python-frontmatter>=1.1",
    "rich>=13.7",
    "pyyaml>=6.0",
]

[project.scripts]
figmaclaw = "figmaclaw.main:cli"

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "basedpyright>=1.14",
    "ruff>=0.4",
    "pre-commit>=3.7",
]

[dependency-groups]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "basedpyright>=1.14",
    "ruff>=0.4",
    "pre-commit>=3.7",
    "respx>=0.22.0",
    "pytest-cov>=7.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "ANN"]
ignore = ["ANN101", "ANN102", "ANN401", "E501"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ANN"]

[tool.basedpyright]
pythonVersion = "3.12"
typeCheckingMode = "standard"
reportMissingImports = true
reportMissingTypeStubs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "smoke: marks tests as smoke tests requiring real credentials (deselect with '-m \"not smoke\"')",
]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["figmaclaw"]

[tool.hatch.build.targets.wheel.force-include]
"figmaclaw/workflows" = "figmaclaw/workflows"
"figmaclaw/skills" = "figmaclaw/skills"
"figmaclaw/prompts" = "figmaclaw/prompts"
